+++ tests/test_report_data_description.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+# tests/test_report_generation.py
+# Copyright (C) 2020 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from rest_framework.test import APIClient
+from django.urls import reverse
+
+
+def test_markdown_description_without_error():
+    client = APIClient()
+    url = reverse('scanreport_data_description')
+    response = client.get(url, format='text/markdown+table')
+    assert response.status_code == 200